move MyParserClass get notification when at home
